+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:name="example.capston.App"
         tools:targetApi="31">
         <activity
-            android:name="com.capston.presentation.MainActivity"
+            android:name="com.capston.presentation.ui.MainActivity"
             android:windowSoftInputMode="adjustPan"
             android:launchMode="singleTop"
             android:exported="true">
