+++ presentation/src/main/java/com/capston/presentation/ui/LectureListScreen.kt
@@ -1,4 +1,4 @@
-package com.capston.presentation
+package com.capston.presentation.ui
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
